Trying to fix crash on 32bit processor: Switch from  to  when decoding oid

+++ Sources/asn1swift/ASN1Serialization.swift
@@ -72,20 +72,20 @@ struct ASN1Deserializer
 	{
 		if contentData.isEmpty { return "" }
 		
-		var oid: [UInt64] = [UInt64]()
+		var oid: [UInt32] = [UInt32]()
 		
 		var shifted: UInt8 = 0x00
-		var value: UInt64 = 0x00
+		var value: UInt32 = 0x00
 		
 		for (i, bit) in contentData.enumerated()
 		{
 			if i == 0
 			{
-				oid.append(UInt64(bit/40))
-				oid.append(UInt64(bit%40))
+				oid.append(UInt32(bit/40))
+				oid.append(UInt32(bit%40))
 			}else if (bit & 0x80) == 0
 			{
-				let v = UInt64((bit & 0x7F) | shifted)
+				let v = UInt32((bit & 0x7F) | shifted)
 				value |= v
 				oid.append(value)
 				
@@ -95,7 +95,7 @@ struct ASN1Deserializer
 			{
 				if value > 0 { value >>= 1 }
 				
-				let v = UInt64(((bit & 0x7F) | shifted) >> 1)
+				let v = UInt32(((bit & 0x7F) | shifted) >> 1)
 				value |= v
 				value <<= 8
